Tweaks

+++ go/service/pprof.go
@@ -32,7 +32,7 @@ func NewPprofHandler(xp rpc.Transporter, g *libkb.GlobalContext) *PprofHandler {
 func (c *PprofHandler) Trace(_ context.Context, arg keybase1.TraceArg) (err error) {
 	sessionID := arg.SessionID
 	traceFile := arg.TraceFile
-	traceDurationSeconds := float64(arg.TraceDurationSeconds)
+	traceDurationSeconds := arg.TraceDurationSeconds
 
 	ctx := engine.Context{
 		LogUI:     c.getLogUI(sessionID),
@@ -48,7 +48,7 @@ func (c *PprofHandler) Trace(_ context.Context, arg keybase1.TraceArg) (err erro
 
 	defer func() {
 		if err != nil {
-			ctx.LogUI.Warning("Failed to trace to %s for %f second(s): %s", traceFile, traceDurationSeconds, err)
+			ctx.LogUI.Warning("Failed to trace to %s for %.2f second(s): %s", traceFile, traceDurationSeconds, err)
 		}
 	}()
 
@@ -63,10 +63,10 @@ func (c *PprofHandler) Trace(_ context.Context, arg keybase1.TraceArg) (err erro
 		return err
 	}
 
-	c.G().Log.Info("Tracing to %s for %f second(s)", traceFile, traceDurationSeconds)
+	c.G().Log.Info("Tracing to %s for %.2f second(s)", traceFile, traceDurationSeconds)
 
 	go func() {
-		time.Sleep(time.Duration(traceDurationSeconds * float64(time.Second)))
+		time.Sleep(time.Duration(float64(traceDurationSeconds) * float64(time.Second)))
 		trace.Stop()
 		close(f)
 		c.G().Log.Info("Tracing to %s done", traceFile)
